Desktop/end to end project/credit score classification/requirements.txt
pandas
numpy 
seaborn
matplotlib
sklearn
dill
flask
flask-cors
gunicorn
PyYAML
pymongo[srv]
dnspython
xgboost
optuna
evidently
-e.